refactor(account): reorganize account-related types and DTOs into a dedicated dto.go file
- Move account-related request and response types from account.go to a new dto.go file for better organization and separation of concerns.
- Maintain existing functionality while improving code clarity and maintainability by structuring account-related data types in a dedicated file.
- Update imports accordingly to reflect the new structure, ensuring seamless integration with the rest of the application.
